pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-assistant"
version = "0.1.0"
description = "A CLI agent to assist with PR creation and management"
readme = "README.md"
authors = [{ name = "Mann Talati", email = "mann@example.com" }]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "PyGithub>=2.1.1",
    "gitpython>=3.1.40",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "google-generativeai>=0.3.0", # Switched to Gemini
]

[project.scripts]
pr-assistant = "pr_assistant.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.1.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
